Move errors to `error.rs`
