+++ compositor_render/src/transformations/builtin.rs
@@ -6,6 +6,7 @@ use compositor_common::scene::{
 use crate::{renderer::texture::NodeTexture, utils::rgba_to_wgpu_color};
 
 pub mod collection;
+pub mod error;
 pub mod node;
 pub mod params;
